requirements.txt
biopython==1.76
ordered-set==4.0.2
pandas==1.1.5
pyfaidx==0.6.4
PyVCF3==1.0.2
packaging==21.3
pysam==0.18.0
pybedtools==0.7.10
